tests: move expect_exception helpers to util.py
This is quite a common pattern.  I wrote a helper
when doing admin_api_auth_test, let's make it available
to use elsewhere.
